Extract align function into it's own file for use by marshallers
