+++ lib/marshall.js
@@ -3,6 +3,7 @@ var assert = require('assert');
 var parseSignature = require('./signature');
 var put = require('put');
 var Marshallers = require('./marshallers');
+var align = require('./align').align;
 
 var marshall = module.exports = function(signature, data, offset) {
   if (typeof(offset) == 'undefined') offset = 0;
@@ -92,15 +93,6 @@ function write(ps, ele, data) {
   }
 }
 
-function align(ps, n) {
-  var pad = n - ps._offset % n;
-  if (pad === 0 || pad === n) return;
-  // TODO: write8(0) in a loop (3 to 7 times here) could be more efficient
-  var padBuff = Buffer(pad);
-  padBuff.fill(0);
-  ps.put(Buffer(padBuff));
-  ps._offset += pad;
-}
 
 var intTypes = ['y', 'n', 'q', 'i', 'u', 'x', 't'];
 var stringTypes = ['g', 'o', 's'];
